change event_num to long long type

+++ prad2det/include/EventData.h
@@ -36,7 +36,7 @@ struct RawEventData {
     int      event_num    = 0;
     uint8_t  trigger_type = 0;   // main trigger (from event tag: tag - 0x80)
     uint32_t trigger      = 0;   // FP trigger bits (multi-bit, from TI master d[5])
-    int64_t  timestamp    = 0;
+    long long  timestamp    = 0;
 
     // HyCal per-channel data
     int          nch = 0;
@@ -78,7 +78,7 @@ struct ReconEventData {
     int      event_num    = 0;
     uint8_t  trigger_type = 0;   // main trigger (from event tag: tag - 0x80)
     uint32_t trigger_bits = 0;   // FP trigger bits (multi-bit, from TI master d[5])
-    int64_t  timestamp    = 0;
+    long long  timestamp    = 0;
 
     // HyCal clusters
     float total_energy = 0.f;
